Complete day 3
Regex day today!

Probably would've done better if I was more familiar with how regex works but hey I made it work in < 15 minutes and I think that's impressive!

I had to modify my starter code a bit to be able to support just a plain file being piped in, but it works well and really didn't take me all that long.